+++ backends-velox/src/test/scala/org/apache/gluten/execution/FunctionsValidateSuite.scala
@@ -25,7 +25,7 @@ import java.nio.file.Files
 
 import scala.collection.JavaConverters._
 
-class FunctionsValidateTest extends WholeStageTransformerSuite {
+class FunctionsValidateSuite extends WholeStageTransformerSuite {
   override protected val resourcePath: String = "/tpch-data-parquet-velox"
   override protected val fileFormat: String = "parquet"
   private var parquetPath: String = _
