+++ backends-velox/src/test/scala/org/apache/gluten/fuzzer/RowToColumnarFuzzer.scala
@@ -17,7 +17,7 @@
 package org.apache.gluten.fuzzer
 
 import org.apache.gluten.execution.RowToVeloxColumnarExec
-import org.apache.gluten.fuzzer.FuzzerTestResult.Successful
+import org.apache.gluten.fuzzer.FuzzerResult.Successful
 import org.apache.gluten.tags.{FuzzerTest, SkipTestTags}
 
 import org.apache.spark.SparkConf
@@ -25,7 +25,7 @@ import org.apache.spark.sql.DataFrame
 
 @FuzzerTest
 @SkipTestTags
-class RowToColumnarFuzzerTest extends FuzzerTestBase {
+class RowToColumnarFuzzer extends FuzzerBase {
 
   override protected def sparkConf: SparkConf = {
     super.sparkConf
